requirements.txt
flask
tensorflow==2.13.0
numpy==1.24.3
pillow
gunicorn
